+++ src/components/fields/effectiveResolution/EffectiveResolution.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TextEntry } from '@ska-telescope/ska-gui-components';
-import { BAND_LOW, FREQUENCY_UNITS, LAB_IS_BOLD, LAB_POSITION } from '../../../utils/constants';
+import { FREQUENCY_UNITS, LAB_IS_BOLD, LAB_POSITION } from '../../../utils/constants';
 import { calculateVelocity, frequencyConversion } from '../../../utils/helpers';
 import { Box } from '@mui/system';
 
@@ -31,8 +31,6 @@ export default function EffectiveResolutionField({
 }: EffectiveResolutionFieldProps) {
   const [effectiveResolution, setEffectiveResolution] = React.useState('');
 
-  const isLow = () => observingBand === BAND_LOW;
-
   const calculateEffectiveResolution = () => {
     const arr = String(spectralResolution).split(' ');
     if (arr.length > 2) {
@@ -40,10 +38,7 @@ export default function EffectiveResolutionField({
       const resolutionUnits = FREQUENCY_UNITS.find(e => e.label === arr[1]).value;
       const freq = frequencyConversion(frequency, frequencyUnits);
       const ave = resolution * spectralAveraging;
-      const velocity = calculateVelocity(
-        frequencyConversion(ave, resolutionUnits) * (isLow() ? 10000 : 10),
-        freq
-      );
+      const velocity = calculateVelocity(frequencyConversion(ave, resolutionUnits) * 10, freq);
       return `${ave.toFixed(2)} ${arr[1]} (${velocity})`;
     } else {
       return '';