+++ CHANGELOG.md
@@ -27,12 +27,18 @@ Future
   - Standardized field sizes
   - Various visual updates and tweaks
 * Observation page
+  - Set frequency and continuum bandwidth units to default when switching band if unmodified
+  - Fixed default continuum bandwidth values missing for band5a and band5b
+  - Added Continuum Bandwidth validation for Mid and Low
+  - Changed Continuum Bandwidth units and frequency units to a disabled dropdown for Low
   - Fixed Sensitivity Calculator results for Mid AA05 observations
   - Fields moved into a self-contained component
       - Effective Resolution
       - Elevation ( minimum for LOW changed to 20 degrees )
       - Spectral Resolution 
       - Robust
+      - Continuum Bandwidth
+      - Bandwidth
 * Data Products
   - Rebalanced the page and also the Add Data Products page to have standard field sizing
 * Utilities